backend/requirements/base.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
Faker==30.8.2